skapat dem olika layers i tiles

+++ src/bobman/Bomb.java
@@ -3,27 +3,27 @@ package bobman;
 public class Bomb 
 {
 	int damage;
-	int xPos;
-	int yPos;
+	double xPos;
+	double yPos;
 	
 	public Bomb(int damage)
 	{
 		this.damage = damage;
 	}
 
-	public int getxPos() {
+	public double getxPos() {
 		return xPos;
 	}
 
-	public void setxPos(int xPos) {
+	public void setxPos(double xPos) {
 		this.xPos = xPos;
 	}
 
-	public int getyPos() {
+	public double getyPos() {
 		return yPos;
 	}
 
-	public void setyPos(int yPos) {
+	public void setyPos(double yPos) {
 		this.yPos = yPos;
 	}
